Conflicts:
	trunk/src/LV2/gx_sceleton.lv2/gx_sceleton.ttl
	trunk/src/LV2/gx_sceleton_stereo.lv2/gx_sceleton.ttl
	trunk/src/plugins/generated/gcb_95.cc

+++ trunk/tools/pluginloader.pyx
@@ -295,6 +295,6 @@ cdef class Plugin:
     def __dealloc__(self):
         del self.varmap
         if self.handle:
-            if self.p:
+            if self.p and self.p[0].delete_instance:
                 self.p[0].delete_instance(self.p)
             dlclose(self.handle)
